+++ react-project/src/App.jsx
@@ -1,6 +1,6 @@
 import Header from "./components/Header";
 import Todo from "./components/Todo";
-import "./App.css";
+import "./App.css"
 
 function App() {
   return (
